fix: correct code style
