+++ videoCourse/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { LogInPageComponent } from './components/course-log-in/log-in-page/log-in-page.component';
-import { AddNewCourseComponent } from './components/course-add-new/add-new-course.component';
+import { LogInPageComponent } from './shared/course-log-in/log-in-page.component';
+import { AddNewCourseComponent } from './shared/course-add-new/add-new-course.component';
 
 
 const routes: Routes = [
