+++ videoCourse/src/app/app.module.ts
@@ -1,10 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import {Router} from '@angular/router';
-
-
 import { AppRoutingModule } from './app-routing.module';
-import {Routes, RouterModule} from '@angular/router';
 import { AppComponent } from './app.component';
 import { CoursePageComponent } from './components/course-page/course-page.component';
 import { CoursePageHeaderComponent } from './components/course-page/course-page-header/course-page-header.component';
@@ -13,12 +10,11 @@ import { CoursePageSearchComponent } from './components/course-page/course-page-
 import { CoursePageListComponent } from './components/course-page/course-page-list/course-page-list.component';
 import { CourseListItemComponent } from './components/course-page/course-page-list/course-list-item/course-list-item.component';
 import { CoursePageFooterComponent } from './components/course-page/course-page-footer/course-page-footer.component';
-import { ItemBorderStyleDirective } from './components/course-page/course-page-list/course-list-item/item-border-style.directive';
-import { CourseDurationPipe } from './components/course-page/course-page-list/course-list-item/course-duration.pipe';
-import { ConfirmationDelItemComponent } from './components/course-page/course-page-list/confirmation-del-item/confirmation-del-item.component';
-import { LoginModule } from './components/course-log-in/login/login.module';
-import { AddNewCourseComponent } from './components/course-add-new/add-new-course.component';
-
+import { ItemBorderStyleDirective } from './core/directives/item-border-style.directive';
+import { CourseDurationPipe } from './shared/pipes/course-duration.pipe';
+import { ConfirmationDelItemComponent } from './shared/confirmation-del-item/confirmation-del-item.component';
+import { LoginModule } from './shared/course-log-in/login.module';
+import { AddNewCourseComponent } from './shared/course-add-new/add-new-course.component';
 
 @NgModule({
   declarations: [
@@ -45,8 +41,6 @@ import { AddNewCourseComponent } from './components/course-add-new/add-new-cours
   bootstrap: [AppComponent]
 })
 export class AppModule { 
-constructor (router: Router) {
-  console.log("Routes:", JSON.stringify(router.config, undefined, 2))
-}
-
+  constructor (router: Router) {
+  }
 }
