--- videoCourse/src/app/components/course-log-in/autorization.service.ts
@@ -1,14 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class AutorizationService {
-
-  constructor() { }
- 
-  public LoginUser (email: string, pass: string) {
-    console.log(email, pass);
-  }
-
-}
