+++ videoCourse/src/app/components/course-page/course-page-breadcrumbs/course-page-breadcrumbs.component.spec.ts
@@ -1,7 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { CoursePageBreadcrumbsComponent } from './course-page-breadcrumbs.component';
-import { Component, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 
 describe('CoursePageBreadcrumbsComponent', () => {
   let component: CoursePageBreadcrumbsComponent;
