+++ videoCourse/src/app/components/course-page/course-page-breadcrumbs/course-page-breadcrumbs.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-course-page-breadcrumbs',
