+++ videoCourse/src/app/components/course-page/course-page-footer/course-page-footer.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-course-page-footer',
@@ -6,5 +6,4 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./course-page-footer.component.scss']
 })
 export class CoursePageFooterComponent {
-
 }
