+++ videoCourse/src/app/components/course-page/course-page-header/course-page-header.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-course-page-header',
@@ -6,5 +6,4 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./course-page-header.component.scss']
 })
 export class CoursePageHeaderComponent {
-
 }
