+++ videoCourse/src/app/components/course-page/course-page-list/course-list-item/course-list-item.component.ts
@@ -1,6 +1,6 @@
 import { Component, Output, Input, EventEmitter, OnInit,  DoCheck } from '@angular/core';
-import { DataCourseService } from '../../data-course.service';
-import {PageListData} from '../../../models/page-list-data';
+import { DataCourseService } from '../../../../core/services/data-course.service';
+import {PageListData} from '../../../../core/models/page-list-data';
 
 
 @Component({
@@ -8,28 +8,22 @@ import {PageListData} from '../../../models/page-list-data';
   templateUrl: './course-list-item.component.html',
   styleUrls: ['./course-list-item.component.scss']
 })
-export class CourseListItemComponent implements OnInit,  DoCheck {
+export class CourseListItemComponent implements OnInit {
   @Input() courseState: string;
   @Output() onDeleteCourseItem = new EventEmitter<number>();
   public courseData: PageListData [];
 
 constructor(private dataCourseService: DataCourseService) { }
 
-ngOnInit() {
-  this.getDataCourse();
-}
+  ngOnInit() {
+    this.getDataCourse();
+  }
 
-ngDoCheck () {
-  this.courseData = this.dataCourseService.getItemList();
+  public getDataCourse() {
+    this.courseData = this.dataCourseService.getItemList();
+  }
 
-}
-
-getDataCourse() {
-  this.courseData = this.dataCourseService.getItemList();
-}
-
-
-public deleteCourseItem(id: number) {
+  public deleteCourseItem(id: number) {
     this.onDeleteCourseItem.emit(id);
-}
+  }
 }
