+++ videoCourse/src/app/components/course-page/course-page-list/course-page-list.component.scss
@@ -1,6 +1,6 @@
 @font-face {
     font-family: Sourcesanspro;
-    src: url(../../../../fonts/sourcesanspro.ttf);
+    src: url(../../../../assets/fonts/sourcesanspro.ttf);
   }
 
 * {
