+++ videoCourse/src/app/shared/confirmation-del-item/confirmation-del-item.component.scss
