+++ videoCourse/src/app/shared/course-add-new/add-new-course.component.scss
@@ -1,6 +1,6 @@
 @font-face {
     font-family: Sourcesanspro;
-    src: url(../../../fonts/sourcesanspro.ttf);
+    src: url(../../../assets/fonts/sourcesanspro.ttf);
   }
 
 * {
