+++ videoCourse/src/app/shared/course-add-new/add-new-course.component.ts
@@ -1,15 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-add-new-course',
   templateUrl: './add-new-course.component.html',
   styleUrls: ['./add-new-course.component.scss']
 })
-export class AddNewCourseComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
+export class AddNewCourseComponent {
 
 }
