+++ videoCourse/src/app/shared/course-log-in/log-in-page.component.scss
@@ -1,6 +1,6 @@
 @font-face {
     font-family: Sourcesanspro;
-    src: url(../../../../fonts/sourcesanspro.ttf);
+    src: url(../../../assets/fonts/sourcesanspro.ttf);
   }
 
 * {
