+++ videoCourse/src/app/shared/course-log-in/login.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { LogInPageComponent } from '../log-in-page/log-in-page.component';
+import { LogInPageComponent } from '../../shared/course-log-in/log-in-page.component';
