+++ videoCourse/src/app/shared/pipes/course-duration.pipe.spec.ts
