+++ videoCourse/src/app/shared/pipes/course-duration.pipe.ts
